Move lots of things around, got modal working for self hosted, moving pieces together.
