+++ components/cloud_subscribe_result_modal/success.tsx
@@ -10,7 +10,7 @@ import {getSubscriptionProduct} from 'mattermost-redux/selectors/entities/cloud'
 
 import PaymentSuccessStandardSvg from 'components/common/svg_images_components/payment_success_standard_svg';
 
-import IconMessage from 'components/cloud_purchase_modal/icon_message';
+import IconMessage from 'components/purchase_modals/icon_message';
 
 import FullScreenModal from 'components/widgets/modals/full_screen_modal';
