+++ components/purchase_modals/icon_message.scss
