+++ packages/mattermost-redux/src/action_types/cloud.ts
@@ -30,4 +30,5 @@ export default keyMirror({
     RECEIVED_SELF_HOSTED_SIGNUP_PROGRESS: null,
 
     CWS_HEALTH_CHECK_OK: null,
+    RECEIVED_SELF_HOSTED_PRODUCTS: null
 });
